R6 Class representing the resampling scheme for Goodness-of-fit-tests for (linear) models

Class is able to resample model fit, i.e. generate a new data set and refit the model to the new data.

Methods

Public methods

GOF_model_resample$new()

GOF_model_resample$resample()

GOF_model_resample$clone()

Method new()

Usage

GOF_model_resample$new(gof_model_simulator, gof_model_trainer)

Arguments

gof_model_simulator

an instance that implements GOF_model_simulator

gof_model_trainer

an instance that implements GOF_model_trainer

Returns

No explicit return

Method resample()

resamples the dependent variable in data and refits model to that new data set

Usage

GOF_model_resample$resample(model, data, y_name)

Arguments

model

fitted model based on data

data

used to fit model

y_name

string specifying the name of the dependent variable

Returns

a resampled version of model

Method clone()

The objects of this class are cloneable with this method.

Usage

GOF_model_resample$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.